requirements.txt
alembic==1.4.2
altgraph==0.17
amqp==2.6.0
amqpstorm==2.7.1
aniso8601==8.0.0
anyjson==0.3.3
appdirs==1.4.4
apscheduler==3.3.1
astropy==3.2.3
atomicwrites==1.4.0
attrs==19.3.0
billiard==3.6.3.0
black==19.10b0
blinker==1.4
Brotli==1.0.7
cachelib==0.1.1
celery==4.4.6
celery-once==3.0.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
concurrent-log-handler==0.9.9
cycler==0.10.0
decorator==4.4.0
dnspython==1.16.0
elasticsearch==7.8.1
email-validator==1.1.1
et-xmlfile==1.0.1
eventlet==0.26.1
flasgger==0.9.5
Flask==1.1.2
Flask-APScheduler==1.9.0
flask-bootstrap==3.3.7.1
Flask-Compress==1.5.0
Flask-Cors==3.0.8
Flask-Excel==0.0.7
Flask-Migrate==2.5.3
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-Session==0.3.2
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
function-scheduling-distributed-framework==7.9
future==0.18.2
gevent==20.6.2
gevent-websocket==0.10.1
gnsq==1.0.1
greenlet==0.4.16
gunicorn==20.0.4
heartpy==1.2.6
hiredis==1.1.0
hrv-analysis==1.0.3
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jsonschema==3.2.0
kafka-python==1.4.6
kiwisolver==1.2.0
kombu==4.6.11
lml==0.0.9
lupa==1.9
lxml==4.5.1
macholib==1.14
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
mongoengine==0.20.0
monotonic==1.5
more-itertools==8.4.0
nb-log==3.4
nolds==0.5.2
numpy==1.19.0
openpyxl==3.0.5
packaging==20.4
pamqp==2.3.0
pandas==1.0.5
pathspec==0.8.0
pefile==2019.4.18
persist-queue==0.5.0
pika==1.1.0
pikav0==0.1.2b0
pikav1==1.0.1b0
Pillow==7.1.2
pluggy==0.13.1
psutil==5.7.2
py==1.9.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.8
pycryptodomex==3.9.8
pyexcel==0.6.4
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xlsx==0.5.5
PyInstaller==3.4
pymongo==3.5.1
PyMuPDF==1.17.5
PyMySQL==0.9.3
pyparsing==2.4.7
pypiwin32==223
pysnooper==0.0.11
pytest==5.4.3
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.13.1
python-socketio==4.6.0
pytz==2020.1
PyWavelets==1.1.1
pywin32==228
pywin32-ctypes==0.2.0
PyYAML==5.3.1
query-string==0.0.0
rabbitpy==2.0.1
redis==3.3.2
redis2==2.10.6
redis3==3.5.2
redlock-py==1.0.8
regex==2020.6.8
reportlab==3.5.44
requests==2.24.0
rocketmq==0.4.4
rsa==4.6
scipy==1.5.0
six==1.15.0
SQLAlchemy==1.3.18
sqlalchemy-utils==0.36.1
texttable==1.6.3
toml==0.10.1
tomorrow3==1.1.0
typed-ast==1.4.1
tzlocal==2.1
ujson==4.0.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.1
xlrd==1.2.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0